update preprocess utils.py
update raster to solve dependence on nuplan-devkit
